Authorization in endpoints in both apps

See merge request Miglou/tqsproject!36

+++ deliveries_server/src/main/java/tqsua/DeliveriesServer/controller/OrderController.java
@@ -20,6 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import tqsua.DeliveriesServer.model.Notification;
@@ -43,32 +44,35 @@ public class OrderController {
     @Autowired
     private NotificationService notificationService;
 
+    private static final String MESSAGE = "message";
+
     private static final Map<String, String> APP_NAMES = Stream.of(new String[][] {
         { "SafeDeliveries", "http://localhost:8081" }, 
       }).collect(Collectors.toMap(data -> data[0], data -> data[1]));
 
 
     //@CrossOrigin(origins = "http://localhost:4200")
-    @GetMapping(path="/orders")
+    @GetMapping(path="/private/orders")
     public ArrayList<Order> getAllOrders() throws IOException, InterruptedException {
+        // TODO:
+        // Verify admin
         return orderService.getAllOrders();
     }
 
     @PostMapping(path="/orders")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<Object> createOrder(@Valid @RequestBody OrderDTO o) throws IOException, InterruptedException {
-        String message = "message";
         HashMap<String, String> response = new HashMap<>();
         if (o.getDeliver_lat() == null || o.getDeliver_lng() == null || o.getPick_up_lat() == null || o.getPick_up_lng() == null) {
-            response.put(message, "Error. Invalid coords.");
+            response.put(MESSAGE, "Error. Invalid coords.");
             return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         }
         if (o.getWeight()<=0) {
-            response.put(message, "Error. Invalid Weight.");
+            response.put(MESSAGE, "Error. Invalid Weight.");
             return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         }
         if (!APP_NAMES.containsKey(o.getApp_name())) {
-            response.put(message, "Error. Invalid App name.");
+            response.put(MESSAGE, "Error. Invalid App name.");
             return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         }
         Order o1 = new Order(0, o.getPick_up_lat(), o.getPick_up_lng(), o.getDeliver_lat(), o.getDeliver_lng(), o.getWeight(), o.getApp_name());
@@ -88,18 +92,34 @@ public class OrderController {
     }
 
 
-    @PostMapping(path="/acceptorder")
+    @PostMapping(path="/private/acceptorder")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<Object> acceptOrder(@RequestParam(name="order_id") long order_id, @RequestParam(name="rider_id") long rider_id) {
+    public ResponseEntity<Object> acceptOrder(Authentication authentication ,@RequestParam(name="order_id") long order_id, @RequestParam(name="rider_id") long rider_id) {
+        String id = authentication.getName();
+
+        if (!id.equals(String.valueOf(rider_id))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put(MESSAGE, "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
+
         notificationService.delete(rider_id);
         Order order = orderService.updateRider(order_id, rider_id);
         riderService.changeStatus(rider_id, "Delivering");
         return new ResponseEntity<>(order, HttpStatus.OK);
     }
 
-    @PostMapping(path="/declineorder")
+    @PostMapping(path="/private/declineorder")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<Object> declineOrder(@RequestParam(name="order_id") long order_id, @RequestParam(name="rider_id") long rider_id) throws IOException, InterruptedException {
+    public ResponseEntity<Object> declineOrder(Authentication authentication ,@RequestParam(name="order_id") long order_id, @RequestParam(name="rider_id") long rider_id) throws IOException, InterruptedException {
+        String id = authentication.getName();
+
+        if (!id.equals(String.valueOf(rider_id))) {
+            HashMap<String, String> response = new HashMap<>();
+            response.put(MESSAGE, "Unauthorized");
+            return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
+        }
+        
         notificationService.delete(rider_id);
 
         Order o = orderService.getOrderById(order_id);
@@ -108,8 +128,8 @@ public class OrderController {
         o.setRefused_riders(refused_riders);
         orderService.saveOrder(o);
         ArrayList<Rider> riders = riderService.getAvailableRiders(o.getWeight());
-        for (Long id: refused_riders) {
-            riders.removeIf(item -> item.getId() == id);
+        for (Long idRider: refused_riders) {
+            riders.removeIf(item -> item.getId() == idRider);
         }
         if (riders.size() != 0) {
             Rider final_rider = getFinalRider(o, riders);
